requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
joblib==1.3.0
pandas==1.5.3
mysql-connector-python==8.0.33
scikit-learn==1.3.0 --prefer-binary
